added icon & splash screen
