+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:recipe_book_flutter/splash_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -8,32 +9,8 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Recipe Book Flutter',
-      home: Home(),
-    );
-  }
-}
-
-class Home extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    final appBar = AppBar(
-      title: Text('Recipe Book'),
-      actions: [
-        IconButton(
-          icon: Icon(Icons.search),
-          onPressed: () {},
-        )
-      ],
-    );
-
-    final body = Center(
-      child: Text('Welcome to the Recipe Book'),
-    );
-
-    return Scaffold(
-      appBar: appBar,
-      body: body,
+      title: 'Recipe Book',
+      home: SplashScreen(),
     );
   }
 }
