+++ pubspec.yaml
@@ -23,11 +23,15 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_launcher_icons:
 
+flutter_icons:
+  image_path: "assets/icon.png"
+  android: true
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  # cupertino_icons: ^1.0.0
 
 dev_dependencies:
   flutter_test:
@@ -45,8 +49,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
+  assets:
+    - assets/icon.png
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
